Lean D-Bus (used only for signals); direct database connection (no IPC); minor changes; missing emit signals
