+++ gawake-cli/meson.build
@@ -1,6 +1,6 @@
 project('gawake-cli', 'c',
           version: '3.1.0',
-    meson_version: '>= 0.62.0',
+    meson_version: '>= 0.59.0',
   default_options: [ 'warning_level=2', 'werror=false', 'c_std=gnu18', ],
 )
 # Note: default template c_std=gnu11
