--- gawake-cli/src/dbus-server/io.github.kelvinnovais.xml
@@ -1,100 +1,0 @@
-<node>
-  <interface name="io.github.kelvinnovais.Database">
-    <!-- METHODS -->
-    <!-- For the methods, data types described on "struct Rules" at gawake_types.h -->
-
-    <!-- Since the first methods are simple, and there isn't a big amount of data, -->
-    <!-- the parameters are obtained separetely -->
-    <method name="AddRule">
-      <arg name="name" direction="in" type="s" />
-      <arg name="hour" direction="in" type="y" />
-      <arg name="minutes" direction="in" type="y" />
-      <arg name="day_0" direction="in" type="b" />
-      <arg name="day_1" direction="in" type="b" />
-      <arg name="day_2" direction="in" type="b" />
-      <arg name="day_3" direction="in" type="b" />
-      <arg name="day_4" direction="in" type="b" />
-      <arg name="day_5" direction="in" type="b" />
-      <arg name="day_6" direction="in" type="b" />
-      <arg name="mode" direction="in" type="y" />
-      <arg name="table" direction="in" type="y" />
-
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="EditRule">
-      <arg name="id" direction="in" type="q" />
-      <arg name="name" direction="in" type="s" />
-      <arg name="hour" direction="in" type="y" />
-      <arg name="minutes" direction="in" type="y" />
-      <arg name="day_0" direction="in" type="b" />
-      <arg name="day_1" direction="in" type="b" />
-      <arg name="day_2" direction="in" type="b" />
-      <arg name="day_3" direction="in" type="b" />
-      <arg name="day_4" direction="in" type="b" />
-      <arg name="day_5" direction="in" type="b" />
-      <arg name="day_6" direction="in" type="b" />
-      <arg name="active" direction="in" type="b" />
-      <arg name="mode" direction="in" type="y" />
-      <arg name="table" direction="in" type="y" />
-
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="DeleteRule">
-      <arg name="id" direction="in" type="q" />
-      <arg name="table" direction="in" type="y" />
-
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="EnableDisableRule">
-      <arg name="id" direction="in" type="q" />
-      <arg name="table" direction="in" type="y" />
-      <arg name="active" direction="in" type="b" />
-
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="QueryRule">
-      <arg name="id" direction="in" type="q" />
-      <arg name="table" direction="in" type="y" />
-
-      <!-- This is a container type: https://dbus.freedesktop.org/doc/dbus-specification.html#container-types -->
-      <arg name="rule" direction="out" type="(qsyybbbbbbbbyy)" /> <!-- If implementing array: qsyyabbyy -->
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="QueryRules">
-      <arg name="table" direction="in" type="y" />
-
-      <arg name="rules" direction="out" type="a(qsyybbbbbbbbyy)" />
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <method name="CancelRule" />
-
-    <method name="Schedule" />
-
-    <method name="CustomSchedule">
-      <arg name="hour" direction="in" type="y" />
-      <arg name="minutes" direction="in" type="y" />
-      <arg name="day" direction="in" type="y" />
-      <arg name="month" direction="in" type="y" />
-      <arg name="year" direction="in" type="q" />
-      <arg name="mode" direction="in" type="y" />
-
-      <arg name="success" direction="out" type="b"/>
-    </method>
-
-    <!-- SIGNALS -->
-    <signal name="DatabaseUpdated" />
-    <signal name="RuleCanceled" />
-    <signal name="ScheduleRequested" />
-  </interface>
-</node>
-
-<!-- Documentation available by running man "gdbus-codegen" -->
-<!-- or at https://manpages.ubuntu.com/manpages/focal/en/man1/gdbus-codegen.1.html -->
-
-<!-- Some help: https://stackoverflow.com/questions/37185735/sample-program-for-gdbus-signals -->
