+++ gawake-cli/src/gawake-cli/meson.build
@@ -1,9 +1,9 @@
 gawake_cli_sources = files(
 	'main.c',
-	'dbus-client.c',
 
 	'../utils/gawake-types.c',
-	'../dbus-server/dbus-server.c',
 	'../utils/debugger.c',
-	'../utils/get-time.c'
+	'../utils/get-time.c',
+	'../utils/validate-rtcwake-args.c',
+	'../utils/database-connection.c'
 )
