+++ gawake-cli/src/gawake-dbus-server/dbus-server.c
@@ -1,0 +1,1696 @@
+/*
+ * This file is generated by gdbus-codegen, do not modify it.
+ *
+ * The license of this code is the same as for the D-Bus interface description
+ * it was derived from. Note that it links to GLib, so must comply with the
+ * LGPL linking clauses.
+ */
+
+#ifdef HAVE_CONFIG_H
+#  include "config.h"
+#endif
+
+#include "dbus-server.h"
+
+#include <string.h>
+#ifdef G_OS_UNIX
+#  include <gio/gunixfdlist.h>
+#endif
+
+#ifdef G_ENABLE_DEBUG
+#define g_marshal_value_peek_boolean(v)  g_value_get_boolean (v)
+#define g_marshal_value_peek_char(v)     g_value_get_schar (v)
+#define g_marshal_value_peek_uchar(v)    g_value_get_uchar (v)
+#define g_marshal_value_peek_int(v)      g_value_get_int (v)
+#define g_marshal_value_peek_uint(v)     g_value_get_uint (v)
+#define g_marshal_value_peek_long(v)     g_value_get_long (v)
+#define g_marshal_value_peek_ulong(v)    g_value_get_ulong (v)
+#define g_marshal_value_peek_int64(v)    g_value_get_int64 (v)
+#define g_marshal_value_peek_uint64(v)   g_value_get_uint64 (v)
+#define g_marshal_value_peek_enum(v)     g_value_get_enum (v)
+#define g_marshal_value_peek_flags(v)    g_value_get_flags (v)
+#define g_marshal_value_peek_float(v)    g_value_get_float (v)
+#define g_marshal_value_peek_double(v)   g_value_get_double (v)
+#define g_marshal_value_peek_string(v)   (char*) g_value_get_string (v)
+#define g_marshal_value_peek_param(v)    g_value_get_param (v)
+#define g_marshal_value_peek_boxed(v)    g_value_get_boxed (v)
+#define g_marshal_value_peek_pointer(v)  g_value_get_pointer (v)
+#define g_marshal_value_peek_object(v)   g_value_get_object (v)
+#define g_marshal_value_peek_variant(v)  g_value_get_variant (v)
+#else /* !G_ENABLE_DEBUG */
+/* WARNING: This code accesses GValues directly, which is UNSUPPORTED API.
+ *          Do not access GValues directly in your code. Instead, use the
+ *          g_value_get_*() functions
+ */
+#define g_marshal_value_peek_boolean(v)  (v)->data[0].v_int
+#define g_marshal_value_peek_char(v)     (v)->data[0].v_int
+#define g_marshal_value_peek_uchar(v)    (v)->data[0].v_uint
+#define g_marshal_value_peek_int(v)      (v)->data[0].v_int
+#define g_marshal_value_peek_uint(v)     (v)->data[0].v_uint
+#define g_marshal_value_peek_long(v)     (v)->data[0].v_long
+#define g_marshal_value_peek_ulong(v)    (v)->data[0].v_ulong
+#define g_marshal_value_peek_int64(v)    (v)->data[0].v_int64
+#define g_marshal_value_peek_uint64(v)   (v)->data[0].v_uint64
+#define g_marshal_value_peek_enum(v)     (v)->data[0].v_long
+#define g_marshal_value_peek_flags(v)    (v)->data[0].v_ulong
+#define g_marshal_value_peek_float(v)    (v)->data[0].v_float
+#define g_marshal_value_peek_double(v)   (v)->data[0].v_double
+#define g_marshal_value_peek_string(v)   (v)->data[0].v_pointer
+#define g_marshal_value_peek_param(v)    (v)->data[0].v_pointer
+#define g_marshal_value_peek_boxed(v)    (v)->data[0].v_pointer
+#define g_marshal_value_peek_pointer(v)  (v)->data[0].v_pointer
+#define g_marshal_value_peek_object(v)   (v)->data[0].v_pointer
+#define g_marshal_value_peek_variant(v)  (v)->data[0].v_pointer
+#endif /* !G_ENABLE_DEBUG */
+
+typedef struct
+{
+  GDBusArgInfo parent_struct;
+  gboolean use_gvariant;
+} _ExtendedGDBusArgInfo;
+
+typedef struct
+{
+  GDBusMethodInfo parent_struct;
+  const gchar *signal_name;
+  gboolean pass_fdlist;
+} _ExtendedGDBusMethodInfo;
+
+typedef struct
+{
+  GDBusSignalInfo parent_struct;
+  const gchar *signal_name;
+} _ExtendedGDBusSignalInfo;
+
+typedef struct
+{
+  GDBusPropertyInfo parent_struct;
+  const gchar *hyphen_name;
+  guint use_gvariant : 1;
+  guint emits_changed_signal : 1;
+} _ExtendedGDBusPropertyInfo;
+
+typedef struct
+{
+  GDBusInterfaceInfo parent_struct;
+  const gchar *hyphen_name;
+} _ExtendedGDBusInterfaceInfo;
+
+typedef struct
+{
+  const _ExtendedGDBusPropertyInfo *info;
+  guint prop_id;
+  GValue orig_value; /* the value before the change */
+} ChangedProperty;
+
+static void
+_changed_property_free (ChangedProperty *data)
+{
+  g_value_unset (&data->orig_value);
+  g_free (data);
+}
+
+static gboolean
+_g_strv_equal0 (gchar **a, gchar **b)
+{
+  gboolean ret = FALSE;
+  guint n;
+  if (a == NULL && b == NULL)
+    {
+      ret = TRUE;
+      goto out;
+    }
+  if (a == NULL || b == NULL)
+    goto out;
+  if (g_strv_length (a) != g_strv_length (b))
+    goto out;
+  for (n = 0; a[n] != NULL; n++)
+    if (g_strcmp0 (a[n], b[n]) != 0)
+      goto out;
+  ret = TRUE;
+out:
+  return ret;
+}
+
+static gboolean
+_g_variant_equal0 (GVariant *a, GVariant *b)
+{
+  gboolean ret = FALSE;
+  if (a == NULL && b == NULL)
+    {
+      ret = TRUE;
+      goto out;
+    }
+  if (a == NULL || b == NULL)
+    goto out;
+  ret = g_variant_equal (a, b);
+out:
+  return ret;
+}
+
+G_GNUC_UNUSED static gboolean
+_g_value_equal (const GValue *a, const GValue *b)
+{
+  gboolean ret = FALSE;
+  g_assert (G_VALUE_TYPE (a) == G_VALUE_TYPE (b));
+  switch (G_VALUE_TYPE (a))
+    {
+      case G_TYPE_BOOLEAN:
+        ret = (g_value_get_boolean (a) == g_value_get_boolean (b));
+        break;
+      case G_TYPE_UCHAR:
+        ret = (g_value_get_uchar (a) == g_value_get_uchar (b));
+        break;
+      case G_TYPE_INT:
+        ret = (g_value_get_int (a) == g_value_get_int (b));
+        break;
+      case G_TYPE_UINT:
+        ret = (g_value_get_uint (a) == g_value_get_uint (b));
+        break;
+      case G_TYPE_INT64:
+        ret = (g_value_get_int64 (a) == g_value_get_int64 (b));
+        break;
+      case G_TYPE_UINT64:
+        ret = (g_value_get_uint64 (a) == g_value_get_uint64 (b));
+        break;
+      case G_TYPE_DOUBLE:
+        {
+          /* Avoid -Wfloat-equal warnings by doing a direct bit compare */
+          gdouble da = g_value_get_double (a);
+          gdouble db = g_value_get_double (b);
+          ret = memcmp (&da, &db, sizeof (gdouble)) == 0;
+        }
+        break;
+      case G_TYPE_STRING:
+        ret = (g_strcmp0 (g_value_get_string (a), g_value_get_string (b)) == 0);
+        break;
+      case G_TYPE_VARIANT:
+        ret = _g_variant_equal0 (g_value_get_variant (a), g_value_get_variant (b));
+        break;
+      default:
+        if (G_VALUE_TYPE (a) == G_TYPE_STRV)
+          ret = _g_strv_equal0 (g_value_get_boxed (a), g_value_get_boxed (b));
+        else
+          g_critical ("_g_value_equal() does not handle type %s", g_type_name (G_VALUE_TYPE (a)));
+        break;
+    }
+  return ret;
+}
+
+static void
+_g_dbus_codegen_marshal_BOOLEAN__OBJECT (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint G_GNUC_UNUSED,
+    void         *marshal_data)
+{
+  typedef gboolean (*_GDbusCodegenMarshalBoolean_ObjectFunc)
+       (void *data1,
+        GDBusMethodInvocation *arg_method_invocation,
+        void *data2);
+  _GDbusCodegenMarshalBoolean_ObjectFunc callback;
+  GCClosure *cc = (GCClosure*) closure;
+  void *data1, *data2;
+  gboolean v_return;
+
+  g_return_if_fail (return_value != NULL);
+  g_return_if_fail (n_param_values == 2);
+
+  if (G_CCLOSURE_SWAP_DATA (closure))
+    {
+      data1 = closure->data;
+      data2 = g_value_peek_pointer (param_values + 0);
+    }
+  else
+    {
+      data1 = g_value_peek_pointer (param_values + 0);
+      data2 = closure->data;
+    }
+
+  callback = (_GDbusCodegenMarshalBoolean_ObjectFunc)
+    (marshal_data ? marshal_data : cc->callback);
+
+  v_return =
+    callback (data1,
+              g_marshal_value_peek_object (param_values + 1),
+              data2);
+
+  g_value_set_boolean (return_value, v_return);
+}
+
+/* ------------------------------------------------------------------------
+ * Code for interface io.github.kelvinnovais.Database
+ * ------------------------------------------------------------------------
+ */
+
+/**
+ * SECTION:GawakeServerDatabase
+ * @title: GawakeServerDatabase
+ * @short_description: Generated C code for the io.github.kelvinnovais.Database D-Bus interface
+ *
+ * This section contains code for working with the <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link> D-Bus interface in C.
+ */
+
+enum
+{
+  GAWAKE_SERVER__DATABASE_DATABASE_UPDATED,
+  GAWAKE_SERVER__DATABASE_RULE_CANCELED,
+  GAWAKE_SERVER__DATABASE_SCHEDULE_REQUESTED,
+};
+
+static unsigned GAWAKE_SERVER__DATABASE_SIGNALS[3] = { 0 };
+
+/* ---- Introspection data for io.github.kelvinnovais.Database ---- */
+
+static const _ExtendedGDBusMethodInfo _gawake_server_database_method_info_update_database =
+{
+  {
+    -1,
+    (gchar *) "UpdateDatabase",
+    NULL,
+    NULL,
+    NULL
+  },
+  "handle-update-database",
+  FALSE
+};
+
+static const _ExtendedGDBusMethodInfo _gawake_server_database_method_info_cancel_rule =
+{
+  {
+    -1,
+    (gchar *) "CancelRule",
+    NULL,
+    NULL,
+    NULL
+  },
+  "handle-cancel-rule",
+  FALSE
+};
+
+static const _ExtendedGDBusMethodInfo _gawake_server_database_method_info_schedule =
+{
+  {
+    -1,
+    (gchar *) "Schedule",
+    NULL,
+    NULL,
+    NULL
+  },
+  "handle-schedule",
+  FALSE
+};
+
+static const GDBusMethodInfo * const _gawake_server_database_method_info_pointers[] =
+{
+  &_gawake_server_database_method_info_update_database.parent_struct,
+  &_gawake_server_database_method_info_cancel_rule.parent_struct,
+  &_gawake_server_database_method_info_schedule.parent_struct,
+  NULL
+};
+
+static const _ExtendedGDBusSignalInfo _gawake_server_database_signal_info_database_updated =
+{
+  {
+    -1,
+    (gchar *) "DatabaseUpdated",
+    NULL,
+    NULL
+  },
+  "database-updated"
+};
+
+static const _ExtendedGDBusSignalInfo _gawake_server_database_signal_info_rule_canceled =
+{
+  {
+    -1,
+    (gchar *) "RuleCanceled",
+    NULL,
+    NULL
+  },
+  "rule-canceled"
+};
+
+static const _ExtendedGDBusSignalInfo _gawake_server_database_signal_info_schedule_requested =
+{
+  {
+    -1,
+    (gchar *) "ScheduleRequested",
+    NULL,
+    NULL
+  },
+  "schedule-requested"
+};
+
+static const GDBusSignalInfo * const _gawake_server_database_signal_info_pointers[] =
+{
+  &_gawake_server_database_signal_info_database_updated.parent_struct,
+  &_gawake_server_database_signal_info_rule_canceled.parent_struct,
+  &_gawake_server_database_signal_info_schedule_requested.parent_struct,
+  NULL
+};
+
+static const _ExtendedGDBusInterfaceInfo _gawake_server_database_interface_info =
+{
+  {
+    -1,
+    (gchar *) "io.github.kelvinnovais.Database",
+    (GDBusMethodInfo **) &_gawake_server_database_method_info_pointers,
+    (GDBusSignalInfo **) &_gawake_server_database_signal_info_pointers,
+    NULL,
+    NULL
+  },
+  "database",
+};
+
+
+/**
+ * gawake_server_database_interface_info:
+ *
+ * Gets a machine-readable description of the <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link> D-Bus interface.
+ *
+ * Returns: (transfer none): A #GDBusInterfaceInfo. Do not free.
+ */
+GDBusInterfaceInfo *
+gawake_server_database_interface_info (void)
+{
+  return (GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct;
+}
+
+/**
+ * gawake_server_database_override_properties:
+ * @klass: The class structure for a #GObject derived class.
+ * @property_id_begin: The property id to assign to the first overridden property.
+ *
+ * Overrides all #GObject properties in the #GawakeServerDatabase interface for a concrete class.
+ * The properties are overridden in the order they are defined.
+ *
+ * Returns: The last property id.
+ */
+guint
+gawake_server_database_override_properties (GObjectClass *klass G_GNUC_UNUSED, guint property_id_begin)
+{
+  return property_id_begin - 1;
+}
+
+
+inline static void
+gawake_server_database_signal_marshal_database_updated (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  g_cclosure_marshal_VOID__VOID (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+inline static void
+gawake_server_database_signal_marshal_rule_canceled (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  g_cclosure_marshal_VOID__VOID (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+inline static void
+gawake_server_database_signal_marshal_schedule_requested (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  g_cclosure_marshal_VOID__VOID (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+inline static void
+gawake_server_database_method_marshal_update_database (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  _g_dbus_codegen_marshal_BOOLEAN__OBJECT (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+inline static void
+gawake_server_database_method_marshal_cancel_rule (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  _g_dbus_codegen_marshal_BOOLEAN__OBJECT (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+inline static void
+gawake_server_database_method_marshal_schedule (
+    GClosure     *closure,
+    GValue       *return_value,
+    unsigned int  n_param_values,
+    const GValue *param_values,
+    void         *invocation_hint,
+    void         *marshal_data)
+{
+  _g_dbus_codegen_marshal_BOOLEAN__OBJECT (closure,
+    return_value, n_param_values, param_values, invocation_hint, marshal_data);
+}
+
+
+/**
+ * GawakeServerDatabase:
+ *
+ * Abstract interface type for the D-Bus interface <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link>.
+ */
+
+/**
+ * GawakeServerDatabaseIface:
+ * @parent_iface: The parent interface.
+ * @handle_cancel_rule: Handler for the #GawakeServerDatabase::handle-cancel-rule signal.
+ * @handle_schedule: Handler for the #GawakeServerDatabase::handle-schedule signal.
+ * @handle_update_database: Handler for the #GawakeServerDatabase::handle-update-database signal.
+ * @database_updated: Handler for the #GawakeServerDatabase::database-updated signal.
+ * @rule_canceled: Handler for the #GawakeServerDatabase::rule-canceled signal.
+ * @schedule_requested: Handler for the #GawakeServerDatabase::schedule-requested signal.
+ *
+ * Virtual table for the D-Bus interface <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link>.
+ */
+
+typedef GawakeServerDatabaseIface GawakeServerDatabaseInterface;
+G_DEFINE_INTERFACE (GawakeServerDatabase, gawake_server_database, G_TYPE_OBJECT)
+
+static void
+gawake_server_database_default_init (GawakeServerDatabaseIface *iface)
+{
+  /* GObject signals for incoming D-Bus method calls: */
+  /**
+   * GawakeServerDatabase::handle-update-database:
+   * @object: A #GawakeServerDatabase.
+   * @invocation: A #GDBusMethodInvocation.
+   *
+   * Signal emitted when a remote caller is invoking the <link linkend="gdbus-method-io-github-kelvinnovais-Database.UpdateDatabase">UpdateDatabase()</link> D-Bus method.
+   *
+   * If a signal handler returns %TRUE, it means the signal handler will handle the invocation (e.g. take a reference to @invocation and eventually call gawake_server_database_complete_update_database() or e.g. g_dbus_method_invocation_return_error() on it) and no other signal handlers will run. If no signal handler handles the invocation, the %G_DBUS_ERROR_UNKNOWN_METHOD error is returned.
+   *
+   * Returns: %G_DBUS_METHOD_INVOCATION_HANDLED or %TRUE if the invocation was handled, %G_DBUS_METHOD_INVOCATION_UNHANDLED or %FALSE to let other signal handlers run.
+   */
+  g_signal_new ("handle-update-database",
+    G_TYPE_FROM_INTERFACE (iface),
+    G_SIGNAL_RUN_LAST,
+    G_STRUCT_OFFSET (GawakeServerDatabaseIface, handle_update_database),
+    g_signal_accumulator_true_handled,
+    NULL,
+      gawake_server_database_method_marshal_update_database,
+    G_TYPE_BOOLEAN,
+    1,
+    G_TYPE_DBUS_METHOD_INVOCATION);
+
+  /**
+   * GawakeServerDatabase::handle-cancel-rule:
+   * @object: A #GawakeServerDatabase.
+   * @invocation: A #GDBusMethodInvocation.
+   *
+   * Signal emitted when a remote caller is invoking the <link linkend="gdbus-method-io-github-kelvinnovais-Database.CancelRule">CancelRule()</link> D-Bus method.
+   *
+   * If a signal handler returns %TRUE, it means the signal handler will handle the invocation (e.g. take a reference to @invocation and eventually call gawake_server_database_complete_cancel_rule() or e.g. g_dbus_method_invocation_return_error() on it) and no other signal handlers will run. If no signal handler handles the invocation, the %G_DBUS_ERROR_UNKNOWN_METHOD error is returned.
+   *
+   * Returns: %G_DBUS_METHOD_INVOCATION_HANDLED or %TRUE if the invocation was handled, %G_DBUS_METHOD_INVOCATION_UNHANDLED or %FALSE to let other signal handlers run.
+   */
+  g_signal_new ("handle-cancel-rule",
+    G_TYPE_FROM_INTERFACE (iface),
+    G_SIGNAL_RUN_LAST,
+    G_STRUCT_OFFSET (GawakeServerDatabaseIface, handle_cancel_rule),
+    g_signal_accumulator_true_handled,
+    NULL,
+      gawake_server_database_method_marshal_cancel_rule,
+    G_TYPE_BOOLEAN,
+    1,
+    G_TYPE_DBUS_METHOD_INVOCATION);
+
+  /**
+   * GawakeServerDatabase::handle-schedule:
+   * @object: A #GawakeServerDatabase.
+   * @invocation: A #GDBusMethodInvocation.
+   *
+   * Signal emitted when a remote caller is invoking the <link linkend="gdbus-method-io-github-kelvinnovais-Database.Schedule">Schedule()</link> D-Bus method.
+   *
+   * If a signal handler returns %TRUE, it means the signal handler will handle the invocation (e.g. take a reference to @invocation and eventually call gawake_server_database_complete_schedule() or e.g. g_dbus_method_invocation_return_error() on it) and no other signal handlers will run. If no signal handler handles the invocation, the %G_DBUS_ERROR_UNKNOWN_METHOD error is returned.
+   *
+   * Returns: %G_DBUS_METHOD_INVOCATION_HANDLED or %TRUE if the invocation was handled, %G_DBUS_METHOD_INVOCATION_UNHANDLED or %FALSE to let other signal handlers run.
+   */
+  g_signal_new ("handle-schedule",
+    G_TYPE_FROM_INTERFACE (iface),
+    G_SIGNAL_RUN_LAST,
+    G_STRUCT_OFFSET (GawakeServerDatabaseIface, handle_schedule),
+    g_signal_accumulator_true_handled,
+    NULL,
+      gawake_server_database_method_marshal_schedule,
+    G_TYPE_BOOLEAN,
+    1,
+    G_TYPE_DBUS_METHOD_INVOCATION);
+
+  /* GObject signals for received D-Bus signals: */
+  /**
+   * GawakeServerDatabase::database-updated:
+   * @object: A #GawakeServerDatabase.
+   *
+   * On the client-side, this signal is emitted whenever the D-Bus signal <link linkend="gdbus-signal-io-github-kelvinnovais-Database.DatabaseUpdated">"DatabaseUpdated"</link> is received.
+   *
+   * On the service-side, this signal can be used with e.g. g_signal_emit_by_name() to make the object emit the D-Bus signal.
+   */
+  GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_DATABASE_UPDATED] =
+    g_signal_new ("database-updated",
+      G_TYPE_FROM_INTERFACE (iface),
+      G_SIGNAL_RUN_LAST,
+      G_STRUCT_OFFSET (GawakeServerDatabaseIface, database_updated),
+      NULL,
+      NULL,
+      gawake_server_database_signal_marshal_database_updated,
+      G_TYPE_NONE,
+      0);
+
+  /**
+   * GawakeServerDatabase::rule-canceled:
+   * @object: A #GawakeServerDatabase.
+   *
+   * On the client-side, this signal is emitted whenever the D-Bus signal <link linkend="gdbus-signal-io-github-kelvinnovais-Database.RuleCanceled">"RuleCanceled"</link> is received.
+   *
+   * On the service-side, this signal can be used with e.g. g_signal_emit_by_name() to make the object emit the D-Bus signal.
+   */
+  GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_RULE_CANCELED] =
+    g_signal_new ("rule-canceled",
+      G_TYPE_FROM_INTERFACE (iface),
+      G_SIGNAL_RUN_LAST,
+      G_STRUCT_OFFSET (GawakeServerDatabaseIface, rule_canceled),
+      NULL,
+      NULL,
+      gawake_server_database_signal_marshal_rule_canceled,
+      G_TYPE_NONE,
+      0);
+
+  /**
+   * GawakeServerDatabase::schedule-requested:
+   * @object: A #GawakeServerDatabase.
+   *
+   * On the client-side, this signal is emitted whenever the D-Bus signal <link linkend="gdbus-signal-io-github-kelvinnovais-Database.ScheduleRequested">"ScheduleRequested"</link> is received.
+   *
+   * On the service-side, this signal can be used with e.g. g_signal_emit_by_name() to make the object emit the D-Bus signal.
+   */
+  GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_SCHEDULE_REQUESTED] =
+    g_signal_new ("schedule-requested",
+      G_TYPE_FROM_INTERFACE (iface),
+      G_SIGNAL_RUN_LAST,
+      G_STRUCT_OFFSET (GawakeServerDatabaseIface, schedule_requested),
+      NULL,
+      NULL,
+      gawake_server_database_signal_marshal_schedule_requested,
+      G_TYPE_NONE,
+      0);
+
+}
+
+/**
+ * gawake_server_database_emit_database_updated:
+ * @object: A #GawakeServerDatabase.
+ *
+ * Emits the <link linkend="gdbus-signal-io-github-kelvinnovais-Database.DatabaseUpdated">"DatabaseUpdated"</link> D-Bus signal.
+ */
+void
+gawake_server_database_emit_database_updated (
+    GawakeServerDatabase *object)
+{
+  g_signal_emit (object, GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_DATABASE_UPDATED], 0);
+}
+
+/**
+ * gawake_server_database_emit_rule_canceled:
+ * @object: A #GawakeServerDatabase.
+ *
+ * Emits the <link linkend="gdbus-signal-io-github-kelvinnovais-Database.RuleCanceled">"RuleCanceled"</link> D-Bus signal.
+ */
+void
+gawake_server_database_emit_rule_canceled (
+    GawakeServerDatabase *object)
+{
+  g_signal_emit (object, GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_RULE_CANCELED], 0);
+}
+
+/**
+ * gawake_server_database_emit_schedule_requested:
+ * @object: A #GawakeServerDatabase.
+ *
+ * Emits the <link linkend="gdbus-signal-io-github-kelvinnovais-Database.ScheduleRequested">"ScheduleRequested"</link> D-Bus signal.
+ */
+void
+gawake_server_database_emit_schedule_requested (
+    GawakeServerDatabase *object)
+{
+  g_signal_emit (object, GAWAKE_SERVER__DATABASE_SIGNALS[GAWAKE_SERVER__DATABASE_SCHEDULE_REQUESTED], 0);
+}
+
+/**
+ * gawake_server_database_call_update_database:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @callback: A #GAsyncReadyCallback to call when the request is satisfied or %NULL.
+ * @user_data: User data to pass to @callback.
+ *
+ * Asynchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.UpdateDatabase">UpdateDatabase()</link> D-Bus method on @proxy.
+ * When the operation is finished, @callback will be invoked in the thread-default main loop of the thread you are calling this method from (see g_main_context_push_thread_default()).
+ * You can then call gawake_server_database_call_update_database_finish() to get the result of the operation.
+ *
+ * See gawake_server_database_call_update_database_sync() for the synchronous, blocking version of this method.
+ */
+void
+gawake_server_database_call_update_database (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GAsyncReadyCallback callback,
+    gpointer user_data)
+{
+  g_dbus_proxy_call (G_DBUS_PROXY (proxy),
+    "UpdateDatabase",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    callback,
+    user_data);
+}
+
+/**
+ * gawake_server_database_call_update_database_finish:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @res: The #GAsyncResult obtained from the #GAsyncReadyCallback passed to gawake_server_database_call_update_database().
+ * @error: Return location for error or %NULL.
+ *
+ * Finishes an operation started with gawake_server_database_call_update_database().
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_update_database_finish (
+    GawakeServerDatabase *proxy,
+    GAsyncResult *res,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_finish (G_DBUS_PROXY (proxy), res, error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_call_update_database_sync:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @error: Return location for error or %NULL.
+ *
+ * Synchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.UpdateDatabase">UpdateDatabase()</link> D-Bus method on @proxy. The calling thread is blocked until a reply is received.
+ *
+ * See gawake_server_database_call_update_database() for the asynchronous version of this method.
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_update_database_sync (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_sync (G_DBUS_PROXY (proxy),
+    "UpdateDatabase",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_call_cancel_rule:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @callback: A #GAsyncReadyCallback to call when the request is satisfied or %NULL.
+ * @user_data: User data to pass to @callback.
+ *
+ * Asynchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.CancelRule">CancelRule()</link> D-Bus method on @proxy.
+ * When the operation is finished, @callback will be invoked in the thread-default main loop of the thread you are calling this method from (see g_main_context_push_thread_default()).
+ * You can then call gawake_server_database_call_cancel_rule_finish() to get the result of the operation.
+ *
+ * See gawake_server_database_call_cancel_rule_sync() for the synchronous, blocking version of this method.
+ */
+void
+gawake_server_database_call_cancel_rule (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GAsyncReadyCallback callback,
+    gpointer user_data)
+{
+  g_dbus_proxy_call (G_DBUS_PROXY (proxy),
+    "CancelRule",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    callback,
+    user_data);
+}
+
+/**
+ * gawake_server_database_call_cancel_rule_finish:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @res: The #GAsyncResult obtained from the #GAsyncReadyCallback passed to gawake_server_database_call_cancel_rule().
+ * @error: Return location for error or %NULL.
+ *
+ * Finishes an operation started with gawake_server_database_call_cancel_rule().
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_cancel_rule_finish (
+    GawakeServerDatabase *proxy,
+    GAsyncResult *res,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_finish (G_DBUS_PROXY (proxy), res, error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_call_cancel_rule_sync:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @error: Return location for error or %NULL.
+ *
+ * Synchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.CancelRule">CancelRule()</link> D-Bus method on @proxy. The calling thread is blocked until a reply is received.
+ *
+ * See gawake_server_database_call_cancel_rule() for the asynchronous version of this method.
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_cancel_rule_sync (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_sync (G_DBUS_PROXY (proxy),
+    "CancelRule",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_call_schedule:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @callback: A #GAsyncReadyCallback to call when the request is satisfied or %NULL.
+ * @user_data: User data to pass to @callback.
+ *
+ * Asynchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.Schedule">Schedule()</link> D-Bus method on @proxy.
+ * When the operation is finished, @callback will be invoked in the thread-default main loop of the thread you are calling this method from (see g_main_context_push_thread_default()).
+ * You can then call gawake_server_database_call_schedule_finish() to get the result of the operation.
+ *
+ * See gawake_server_database_call_schedule_sync() for the synchronous, blocking version of this method.
+ */
+void
+gawake_server_database_call_schedule (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GAsyncReadyCallback callback,
+    gpointer user_data)
+{
+  g_dbus_proxy_call (G_DBUS_PROXY (proxy),
+    "Schedule",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    callback,
+    user_data);
+}
+
+/**
+ * gawake_server_database_call_schedule_finish:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @res: The #GAsyncResult obtained from the #GAsyncReadyCallback passed to gawake_server_database_call_schedule().
+ * @error: Return location for error or %NULL.
+ *
+ * Finishes an operation started with gawake_server_database_call_schedule().
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_schedule_finish (
+    GawakeServerDatabase *proxy,
+    GAsyncResult *res,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_finish (G_DBUS_PROXY (proxy), res, error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_call_schedule_sync:
+ * @proxy: A #GawakeServerDatabaseProxy.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @error: Return location for error or %NULL.
+ *
+ * Synchronously invokes the <link linkend="gdbus-method-io-github-kelvinnovais-Database.Schedule">Schedule()</link> D-Bus method on @proxy. The calling thread is blocked until a reply is received.
+ *
+ * See gawake_server_database_call_schedule() for the asynchronous version of this method.
+ *
+ * Returns: (skip): %TRUE if the call succeeded, %FALSE if @error is set.
+ */
+gboolean
+gawake_server_database_call_schedule_sync (
+    GawakeServerDatabase *proxy,
+    GCancellable *cancellable,
+    GError **error)
+{
+  GVariant *_ret;
+  _ret = g_dbus_proxy_call_sync (G_DBUS_PROXY (proxy),
+    "Schedule",
+    g_variant_new ("()"),
+    G_DBUS_CALL_FLAGS_NONE,
+    -1,
+    cancellable,
+    error);
+  if (_ret == NULL)
+    goto _out;
+  g_variant_get (_ret,
+                 "()");
+  g_variant_unref (_ret);
+_out:
+  return _ret != NULL;
+}
+
+/**
+ * gawake_server_database_complete_update_database:
+ * @object: A #GawakeServerDatabase.
+ * @invocation: (transfer full): A #GDBusMethodInvocation.
+ *
+ * Helper function used in service implementations to finish handling invocations of the <link linkend="gdbus-method-io-github-kelvinnovais-Database.UpdateDatabase">UpdateDatabase()</link> D-Bus method. If you instead want to finish handling an invocation by returning an error, use g_dbus_method_invocation_return_error() or similar.
+ *
+ * This method will free @invocation, you cannot use it afterwards.
+ */
+void
+gawake_server_database_complete_update_database (
+    GawakeServerDatabase *object G_GNUC_UNUSED,
+    GDBusMethodInvocation *invocation)
+{
+  g_dbus_method_invocation_return_value (invocation,
+    g_variant_new ("()"));
+}
+
+/**
+ * gawake_server_database_complete_cancel_rule:
+ * @object: A #GawakeServerDatabase.
+ * @invocation: (transfer full): A #GDBusMethodInvocation.
+ *
+ * Helper function used in service implementations to finish handling invocations of the <link linkend="gdbus-method-io-github-kelvinnovais-Database.CancelRule">CancelRule()</link> D-Bus method. If you instead want to finish handling an invocation by returning an error, use g_dbus_method_invocation_return_error() or similar.
+ *
+ * This method will free @invocation, you cannot use it afterwards.
+ */
+void
+gawake_server_database_complete_cancel_rule (
+    GawakeServerDatabase *object G_GNUC_UNUSED,
+    GDBusMethodInvocation *invocation)
+{
+  g_dbus_method_invocation_return_value (invocation,
+    g_variant_new ("()"));
+}
+
+/**
+ * gawake_server_database_complete_schedule:
+ * @object: A #GawakeServerDatabase.
+ * @invocation: (transfer full): A #GDBusMethodInvocation.
+ *
+ * Helper function used in service implementations to finish handling invocations of the <link linkend="gdbus-method-io-github-kelvinnovais-Database.Schedule">Schedule()</link> D-Bus method. If you instead want to finish handling an invocation by returning an error, use g_dbus_method_invocation_return_error() or similar.
+ *
+ * This method will free @invocation, you cannot use it afterwards.
+ */
+void
+gawake_server_database_complete_schedule (
+    GawakeServerDatabase *object G_GNUC_UNUSED,
+    GDBusMethodInvocation *invocation)
+{
+  g_dbus_method_invocation_return_value (invocation,
+    g_variant_new ("()"));
+}
+
+/* ------------------------------------------------------------------------ */
+
+/**
+ * GawakeServerDatabaseProxy:
+ *
+ * The #GawakeServerDatabaseProxy structure contains only private data and should only be accessed using the provided API.
+ */
+
+/**
+ * GawakeServerDatabaseProxyClass:
+ * @parent_class: The parent class.
+ *
+ * Class structure for #GawakeServerDatabaseProxy.
+ */
+
+struct _GawakeServerDatabaseProxyPrivate
+{
+  GData *qdata;
+};
+
+static void gawake_server_database_proxy_iface_init (GawakeServerDatabaseIface *iface);
+
+#if GLIB_VERSION_MAX_ALLOWED >= GLIB_VERSION_2_38
+G_DEFINE_TYPE_WITH_CODE (GawakeServerDatabaseProxy, gawake_server_database_proxy, G_TYPE_DBUS_PROXY,
+                         G_ADD_PRIVATE (GawakeServerDatabaseProxy)
+                         G_IMPLEMENT_INTERFACE (GAWAKE_SERVER_TYPE_DATABASE, gawake_server_database_proxy_iface_init))
+
+#else
+G_DEFINE_TYPE_WITH_CODE (GawakeServerDatabaseProxy, gawake_server_database_proxy, G_TYPE_DBUS_PROXY,
+                         G_IMPLEMENT_INTERFACE (GAWAKE_SERVER_TYPE_DATABASE, gawake_server_database_proxy_iface_init))
+
+#endif
+static void
+gawake_server_database_proxy_finalize (GObject *object)
+{
+  GawakeServerDatabaseProxy *proxy = GAWAKE_SERVER_DATABASE_PROXY (object);
+  g_datalist_clear (&proxy->priv->qdata);
+  G_OBJECT_CLASS (gawake_server_database_proxy_parent_class)->finalize (object);
+}
+
+static void
+gawake_server_database_proxy_get_property (GObject      *object G_GNUC_UNUSED,
+  guint         prop_id G_GNUC_UNUSED,
+  GValue       *value G_GNUC_UNUSED,
+  GParamSpec   *pspec G_GNUC_UNUSED)
+{
+}
+
+static void
+gawake_server_database_proxy_set_property (GObject      *object G_GNUC_UNUSED,
+  guint         prop_id G_GNUC_UNUSED,
+  const GValue *value G_GNUC_UNUSED,
+  GParamSpec   *pspec G_GNUC_UNUSED)
+{
+}
+
+static void
+gawake_server_database_proxy_g_signal (GDBusProxy *proxy,
+  const gchar *sender_name G_GNUC_UNUSED,
+  const gchar *signal_name,
+  GVariant *parameters)
+{
+  _ExtendedGDBusSignalInfo *info;
+  GVariantIter iter;
+  GVariant *child;
+  GValue *paramv;
+  gsize num_params;
+  gsize n;
+  guint signal_id;
+  info = (_ExtendedGDBusSignalInfo *) g_dbus_interface_info_lookup_signal ((GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct, signal_name);
+  if (info == NULL)
+    return;
+  num_params = g_variant_n_children (parameters);
+  paramv = g_new0 (GValue, num_params + 1);
+  g_value_init (&paramv[0], GAWAKE_SERVER_TYPE_DATABASE);
+  g_value_set_object (&paramv[0], proxy);
+  g_variant_iter_init (&iter, parameters);
+  n = 1;
+  while ((child = g_variant_iter_next_value (&iter)) != NULL)
+    {
+      _ExtendedGDBusArgInfo *arg_info = (_ExtendedGDBusArgInfo *) info->parent_struct.args[n - 1];
+      if (arg_info->use_gvariant)
+        {
+          g_value_init (&paramv[n], G_TYPE_VARIANT);
+          g_value_set_variant (&paramv[n], child);
+          n++;
+        }
+      else
+        g_dbus_gvariant_to_gvalue (child, &paramv[n++]);
+      g_variant_unref (child);
+    }
+  signal_id = g_signal_lookup (info->signal_name, GAWAKE_SERVER_TYPE_DATABASE);
+  g_signal_emitv (paramv, signal_id, 0, NULL);
+  for (n = 0; n < num_params + 1; n++)
+    g_value_unset (&paramv[n]);
+  g_free (paramv);
+}
+
+static void
+gawake_server_database_proxy_g_properties_changed (GDBusProxy *_proxy,
+  GVariant *changed_properties,
+  const gchar *const *invalidated_properties)
+{
+  GawakeServerDatabaseProxy *proxy = GAWAKE_SERVER_DATABASE_PROXY (_proxy);
+  guint n;
+  const gchar *key;
+  GVariantIter *iter;
+  _ExtendedGDBusPropertyInfo *info;
+  g_variant_get (changed_properties, "a{sv}", &iter);
+  while (g_variant_iter_next (iter, "{&sv}", &key, NULL))
+    {
+      info = (_ExtendedGDBusPropertyInfo *) g_dbus_interface_info_lookup_property ((GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct, key);
+      g_datalist_remove_data (&proxy->priv->qdata, key);
+      if (info != NULL)
+        g_object_notify (G_OBJECT (proxy), info->hyphen_name);
+    }
+  g_variant_iter_free (iter);
+  for (n = 0; invalidated_properties[n] != NULL; n++)
+    {
+      info = (_ExtendedGDBusPropertyInfo *) g_dbus_interface_info_lookup_property ((GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct, invalidated_properties[n]);
+      g_datalist_remove_data (&proxy->priv->qdata, invalidated_properties[n]);
+      if (info != NULL)
+        g_object_notify (G_OBJECT (proxy), info->hyphen_name);
+    }
+}
+
+static void
+gawake_server_database_proxy_init (GawakeServerDatabaseProxy *proxy)
+{
+#if GLIB_VERSION_MAX_ALLOWED >= GLIB_VERSION_2_38
+  proxy->priv = gawake_server_database_proxy_get_instance_private (proxy);
+#else
+  proxy->priv = G_TYPE_INSTANCE_GET_PRIVATE (proxy, GAWAKE_SERVER_TYPE_DATABASE_PROXY, GawakeServerDatabaseProxyPrivate);
+#endif
+
+  g_dbus_proxy_set_interface_info (G_DBUS_PROXY (proxy), gawake_server_database_interface_info ());
+}
+
+static void
+gawake_server_database_proxy_class_init (GawakeServerDatabaseProxyClass *klass)
+{
+  GObjectClass *gobject_class;
+  GDBusProxyClass *proxy_class;
+
+  gobject_class = G_OBJECT_CLASS (klass);
+  gobject_class->finalize     = gawake_server_database_proxy_finalize;
+  gobject_class->get_property = gawake_server_database_proxy_get_property;
+  gobject_class->set_property = gawake_server_database_proxy_set_property;
+
+  proxy_class = G_DBUS_PROXY_CLASS (klass);
+  proxy_class->g_signal = gawake_server_database_proxy_g_signal;
+  proxy_class->g_properties_changed = gawake_server_database_proxy_g_properties_changed;
+
+#if GLIB_VERSION_MAX_ALLOWED < GLIB_VERSION_2_38
+  g_type_class_add_private (klass, sizeof (GawakeServerDatabaseProxyPrivate));
+#endif
+}
+
+static void
+gawake_server_database_proxy_iface_init (GawakeServerDatabaseIface *iface G_GNUC_UNUSED)
+{
+}
+
+/**
+ * gawake_server_database_proxy_new:
+ * @connection: A #GDBusConnection.
+ * @flags: Flags from the #GDBusProxyFlags enumeration.
+ * @name: (nullable): A bus name (well-known or unique) or %NULL if @connection is not a message bus connection.
+ * @object_path: An object path.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @callback: A #GAsyncReadyCallback to call when the request is satisfied.
+ * @user_data: User data to pass to @callback.
+ *
+ * Asynchronously creates a proxy for the D-Bus interface <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link>. See g_dbus_proxy_new() for more details.
+ *
+ * When the operation is finished, @callback will be invoked in the thread-default main loop of the thread you are calling this method from (see g_main_context_push_thread_default()).
+ * You can then call gawake_server_database_proxy_new_finish() to get the result of the operation.
+ *
+ * See gawake_server_database_proxy_new_sync() for the synchronous, blocking version of this constructor.
+ */
+void
+gawake_server_database_proxy_new (
+    GDBusConnection     *connection,
+    GDBusProxyFlags      flags,
+    const gchar         *name,
+    const gchar         *object_path,
+    GCancellable        *cancellable,
+    GAsyncReadyCallback  callback,
+    gpointer             user_data)
+{
+  g_async_initable_new_async (GAWAKE_SERVER_TYPE_DATABASE_PROXY, G_PRIORITY_DEFAULT, cancellable, callback, user_data, "g-flags", flags, "g-name", name, "g-connection", connection, "g-object-path", object_path, "g-interface-name", "io.github.kelvinnovais.Database", NULL);
+}
+
+/**
+ * gawake_server_database_proxy_new_finish:
+ * @res: The #GAsyncResult obtained from the #GAsyncReadyCallback passed to gawake_server_database_proxy_new().
+ * @error: Return location for error or %NULL
+ *
+ * Finishes an operation started with gawake_server_database_proxy_new().
+ *
+ * Returns: (transfer full) (type GawakeServerDatabaseProxy): The constructed proxy object or %NULL if @error is set.
+ */
+GawakeServerDatabase *
+gawake_server_database_proxy_new_finish (
+    GAsyncResult        *res,
+    GError             **error)
+{
+  GObject *ret;
+  GObject *source_object;
+  source_object = g_async_result_get_source_object (res);
+  ret = g_async_initable_new_finish (G_ASYNC_INITABLE (source_object), res, error);
+  g_object_unref (source_object);
+  if (ret != NULL)
+    return GAWAKE_SERVER_DATABASE (ret);
+  else
+    return NULL;
+}
+
+/**
+ * gawake_server_database_proxy_new_sync:
+ * @connection: A #GDBusConnection.
+ * @flags: Flags from the #GDBusProxyFlags enumeration.
+ * @name: (nullable): A bus name (well-known or unique) or %NULL if @connection is not a message bus connection.
+ * @object_path: An object path.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @error: Return location for error or %NULL
+ *
+ * Synchronously creates a proxy for the D-Bus interface <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link>. See g_dbus_proxy_new_sync() for more details.
+ *
+ * The calling thread is blocked until a reply is received.
+ *
+ * See gawake_server_database_proxy_new() for the asynchronous version of this constructor.
+ *
+ * Returns: (transfer full) (type GawakeServerDatabaseProxy): The constructed proxy object or %NULL if @error is set.
+ */
+GawakeServerDatabase *
+gawake_server_database_proxy_new_sync (
+    GDBusConnection     *connection,
+    GDBusProxyFlags      flags,
+    const gchar         *name,
+    const gchar         *object_path,
+    GCancellable        *cancellable,
+    GError             **error)
+{
+  GInitable *ret;
+  ret = g_initable_new (GAWAKE_SERVER_TYPE_DATABASE_PROXY, cancellable, error, "g-flags", flags, "g-name", name, "g-connection", connection, "g-object-path", object_path, "g-interface-name", "io.github.kelvinnovais.Database", NULL);
+  if (ret != NULL)
+    return GAWAKE_SERVER_DATABASE (ret);
+  else
+    return NULL;
+}
+
+
+/**
+ * gawake_server_database_proxy_new_for_bus:
+ * @bus_type: A #GBusType.
+ * @flags: Flags from the #GDBusProxyFlags enumeration.
+ * @name: A bus name (well-known or unique).
+ * @object_path: An object path.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @callback: A #GAsyncReadyCallback to call when the request is satisfied.
+ * @user_data: User data to pass to @callback.
+ *
+ * Like gawake_server_database_proxy_new() but takes a #GBusType instead of a #GDBusConnection.
+ *
+ * When the operation is finished, @callback will be invoked in the thread-default main loop of the thread you are calling this method from (see g_main_context_push_thread_default()).
+ * You can then call gawake_server_database_proxy_new_for_bus_finish() to get the result of the operation.
+ *
+ * See gawake_server_database_proxy_new_for_bus_sync() for the synchronous, blocking version of this constructor.
+ */
+void
+gawake_server_database_proxy_new_for_bus (
+    GBusType             bus_type,
+    GDBusProxyFlags      flags,
+    const gchar         *name,
+    const gchar         *object_path,
+    GCancellable        *cancellable,
+    GAsyncReadyCallback  callback,
+    gpointer             user_data)
+{
+  g_async_initable_new_async (GAWAKE_SERVER_TYPE_DATABASE_PROXY, G_PRIORITY_DEFAULT, cancellable, callback, user_data, "g-flags", flags, "g-name", name, "g-bus-type", bus_type, "g-object-path", object_path, "g-interface-name", "io.github.kelvinnovais.Database", NULL);
+}
+
+/**
+ * gawake_server_database_proxy_new_for_bus_finish:
+ * @res: The #GAsyncResult obtained from the #GAsyncReadyCallback passed to gawake_server_database_proxy_new_for_bus().
+ * @error: Return location for error or %NULL
+ *
+ * Finishes an operation started with gawake_server_database_proxy_new_for_bus().
+ *
+ * Returns: (transfer full) (type GawakeServerDatabaseProxy): The constructed proxy object or %NULL if @error is set.
+ */
+GawakeServerDatabase *
+gawake_server_database_proxy_new_for_bus_finish (
+    GAsyncResult        *res,
+    GError             **error)
+{
+  GObject *ret;
+  GObject *source_object;
+  source_object = g_async_result_get_source_object (res);
+  ret = g_async_initable_new_finish (G_ASYNC_INITABLE (source_object), res, error);
+  g_object_unref (source_object);
+  if (ret != NULL)
+    return GAWAKE_SERVER_DATABASE (ret);
+  else
+    return NULL;
+}
+
+/**
+ * gawake_server_database_proxy_new_for_bus_sync:
+ * @bus_type: A #GBusType.
+ * @flags: Flags from the #GDBusProxyFlags enumeration.
+ * @name: A bus name (well-known or unique).
+ * @object_path: An object path.
+ * @cancellable: (nullable): A #GCancellable or %NULL.
+ * @error: Return location for error or %NULL
+ *
+ * Like gawake_server_database_proxy_new_sync() but takes a #GBusType instead of a #GDBusConnection.
+ *
+ * The calling thread is blocked until a reply is received.
+ *
+ * See gawake_server_database_proxy_new_for_bus() for the asynchronous version of this constructor.
+ *
+ * Returns: (transfer full) (type GawakeServerDatabaseProxy): The constructed proxy object or %NULL if @error is set.
+ */
+GawakeServerDatabase *
+gawake_server_database_proxy_new_for_bus_sync (
+    GBusType             bus_type,
+    GDBusProxyFlags      flags,
+    const gchar         *name,
+    const gchar         *object_path,
+    GCancellable        *cancellable,
+    GError             **error)
+{
+  GInitable *ret;
+  ret = g_initable_new (GAWAKE_SERVER_TYPE_DATABASE_PROXY, cancellable, error, "g-flags", flags, "g-name", name, "g-bus-type", bus_type, "g-object-path", object_path, "g-interface-name", "io.github.kelvinnovais.Database", NULL);
+  if (ret != NULL)
+    return GAWAKE_SERVER_DATABASE (ret);
+  else
+    return NULL;
+}
+
+
+/* ------------------------------------------------------------------------ */
+
+/**
+ * GawakeServerDatabaseSkeleton:
+ *
+ * The #GawakeServerDatabaseSkeleton structure contains only private data and should only be accessed using the provided API.
+ */
+
+/**
+ * GawakeServerDatabaseSkeletonClass:
+ * @parent_class: The parent class.
+ *
+ * Class structure for #GawakeServerDatabaseSkeleton.
+ */
+
+struct _GawakeServerDatabaseSkeletonPrivate
+{
+  GValue *properties;
+  GList *changed_properties;
+  GSource *changed_properties_idle_source;
+  GMainContext *context;
+  GMutex lock;
+};
+
+static void
+_gawake_server_database_skeleton_handle_method_call (
+  GDBusConnection *connection G_GNUC_UNUSED,
+  const gchar *sender G_GNUC_UNUSED,
+  const gchar *object_path G_GNUC_UNUSED,
+  const gchar *interface_name,
+  const gchar *method_name,
+  GVariant *parameters,
+  GDBusMethodInvocation *invocation,
+  gpointer user_data)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (user_data);
+  _ExtendedGDBusMethodInfo *info;
+  GVariantIter iter;
+  GVariant *child;
+  GValue *paramv;
+  gsize num_params;
+  guint num_extra;
+  gsize n;
+  guint signal_id;
+  GValue return_value = G_VALUE_INIT;
+  info = (_ExtendedGDBusMethodInfo *) g_dbus_method_invocation_get_method_info (invocation);
+  g_assert (info != NULL);
+  num_params = g_variant_n_children (parameters);
+  num_extra = info->pass_fdlist ? 3 : 2;  paramv = g_new0 (GValue, num_params + num_extra);
+  n = 0;
+  g_value_init (&paramv[n], GAWAKE_SERVER_TYPE_DATABASE);
+  g_value_set_object (&paramv[n++], skeleton);
+  g_value_init (&paramv[n], G_TYPE_DBUS_METHOD_INVOCATION);
+  g_value_set_object (&paramv[n++], invocation);
+  if (info->pass_fdlist)
+    {
+#ifdef G_OS_UNIX
+      g_value_init (&paramv[n], G_TYPE_UNIX_FD_LIST);
+      g_value_set_object (&paramv[n++], g_dbus_message_get_unix_fd_list (g_dbus_method_invocation_get_message (invocation)));
+#else
+      g_assert_not_reached ();
+#endif
+    }
+  g_variant_iter_init (&iter, parameters);
+  while ((child = g_variant_iter_next_value (&iter)) != NULL)
+    {
+      _ExtendedGDBusArgInfo *arg_info = (_ExtendedGDBusArgInfo *) info->parent_struct.in_args[n - num_extra];
+      if (arg_info->use_gvariant)
+        {
+          g_value_init (&paramv[n], G_TYPE_VARIANT);
+          g_value_set_variant (&paramv[n], child);
+          n++;
+        }
+      else
+        g_dbus_gvariant_to_gvalue (child, &paramv[n++]);
+      g_variant_unref (child);
+    }
+  signal_id = g_signal_lookup (info->signal_name, GAWAKE_SERVER_TYPE_DATABASE);
+  g_value_init (&return_value, G_TYPE_BOOLEAN);
+  g_signal_emitv (paramv, signal_id, 0, &return_value);
+  if (!g_value_get_boolean (&return_value))
+    g_dbus_method_invocation_return_error (invocation, G_DBUS_ERROR, G_DBUS_ERROR_UNKNOWN_METHOD, "Method %s is not implemented on interface %s", method_name, interface_name);
+  g_value_unset (&return_value);
+  for (n = 0; n < num_params + num_extra; n++)
+    g_value_unset (&paramv[n]);
+  g_free (paramv);
+}
+
+static GVariant *
+_gawake_server_database_skeleton_handle_get_property (
+  GDBusConnection *connection G_GNUC_UNUSED,
+  const gchar *sender G_GNUC_UNUSED,
+  const gchar *object_path G_GNUC_UNUSED,
+  const gchar *interface_name G_GNUC_UNUSED,
+  const gchar *property_name,
+  GError **error,
+  gpointer user_data)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (user_data);
+  GValue value = G_VALUE_INIT;
+  GParamSpec *pspec;
+  _ExtendedGDBusPropertyInfo *info;
+  GVariant *ret;
+  ret = NULL;
+  info = (_ExtendedGDBusPropertyInfo *) g_dbus_interface_info_lookup_property ((GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct, property_name);
+  g_assert (info != NULL);
+  pspec = g_object_class_find_property (G_OBJECT_GET_CLASS (skeleton), info->hyphen_name);
+  if (pspec == NULL)
+    {
+      g_set_error (error, G_DBUS_ERROR, G_DBUS_ERROR_INVALID_ARGS, "No property with name %s", property_name);
+    }
+  else
+    {
+      g_value_init (&value, pspec->value_type);
+      g_object_get_property (G_OBJECT (skeleton), info->hyphen_name, &value);
+      ret = g_dbus_gvalue_to_gvariant (&value, G_VARIANT_TYPE (info->parent_struct.signature));
+      g_value_unset (&value);
+    }
+  return ret;
+}
+
+static gboolean
+_gawake_server_database_skeleton_handle_set_property (
+  GDBusConnection *connection G_GNUC_UNUSED,
+  const gchar *sender G_GNUC_UNUSED,
+  const gchar *object_path G_GNUC_UNUSED,
+  const gchar *interface_name G_GNUC_UNUSED,
+  const gchar *property_name,
+  GVariant *variant,
+  GError **error,
+  gpointer user_data)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (user_data);
+  GValue value = G_VALUE_INIT;
+  GParamSpec *pspec;
+  _ExtendedGDBusPropertyInfo *info;
+  gboolean ret;
+  ret = FALSE;
+  info = (_ExtendedGDBusPropertyInfo *) g_dbus_interface_info_lookup_property ((GDBusInterfaceInfo *) &_gawake_server_database_interface_info.parent_struct, property_name);
+  g_assert (info != NULL);
+  pspec = g_object_class_find_property (G_OBJECT_GET_CLASS (skeleton), info->hyphen_name);
+  if (pspec == NULL)
+    {
+      g_set_error (error, G_DBUS_ERROR, G_DBUS_ERROR_INVALID_ARGS, "No property with name %s", property_name);
+    }
+  else
+    {
+      if (info->use_gvariant)
+        g_value_set_variant (&value, variant);
+      else
+        g_dbus_gvariant_to_gvalue (variant, &value);
+      g_object_set_property (G_OBJECT (skeleton), info->hyphen_name, &value);
+      g_value_unset (&value);
+      ret = TRUE;
+    }
+  return ret;
+}
+
+static const GDBusInterfaceVTable _gawake_server_database_skeleton_vtable =
+{
+  _gawake_server_database_skeleton_handle_method_call,
+  _gawake_server_database_skeleton_handle_get_property,
+  _gawake_server_database_skeleton_handle_set_property,
+  {NULL}
+};
+
+static GDBusInterfaceInfo *
+gawake_server_database_skeleton_dbus_interface_get_info (GDBusInterfaceSkeleton *skeleton G_GNUC_UNUSED)
+{
+  return gawake_server_database_interface_info ();
+}
+
+static GDBusInterfaceVTable *
+gawake_server_database_skeleton_dbus_interface_get_vtable (GDBusInterfaceSkeleton *skeleton G_GNUC_UNUSED)
+{
+  return (GDBusInterfaceVTable *) &_gawake_server_database_skeleton_vtable;
+}
+
+static GVariant *
+gawake_server_database_skeleton_dbus_interface_get_properties (GDBusInterfaceSkeleton *_skeleton)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (_skeleton);
+
+  GVariantBuilder builder;
+  guint n;
+  g_variant_builder_init (&builder, G_VARIANT_TYPE ("a{sv}"));
+  if (_gawake_server_database_interface_info.parent_struct.properties == NULL)
+    goto out;
+  for (n = 0; _gawake_server_database_interface_info.parent_struct.properties[n] != NULL; n++)
+    {
+      GDBusPropertyInfo *info = _gawake_server_database_interface_info.parent_struct.properties[n];
+      if (info->flags & G_DBUS_PROPERTY_INFO_FLAGS_READABLE)
+        {
+          GVariant *value;
+          value = _gawake_server_database_skeleton_handle_get_property (g_dbus_interface_skeleton_get_connection (G_DBUS_INTERFACE_SKELETON (skeleton)), NULL, g_dbus_interface_skeleton_get_object_path (G_DBUS_INTERFACE_SKELETON (skeleton)), "io.github.kelvinnovais.Database", info->name, NULL, skeleton);
+          if (value != NULL)
+            {
+              g_variant_take_ref (value);
+              g_variant_builder_add (&builder, "{sv}", info->name, value);
+              g_variant_unref (value);
+            }
+        }
+    }
+out:
+  return g_variant_builder_end (&builder);
+}
+
+static void
+gawake_server_database_skeleton_dbus_interface_flush (GDBusInterfaceSkeleton *_skeleton G_GNUC_UNUSED)
+{
+}
+
+static void
+_gawake_server_database_on_signal_database_updated (
+    GawakeServerDatabase *object)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (object);
+
+  GList      *connections, *l;
+  GVariant   *signal_variant;
+  connections = g_dbus_interface_skeleton_get_connections (G_DBUS_INTERFACE_SKELETON (skeleton));
+
+  signal_variant = g_variant_ref_sink (g_variant_new ("()"));
+  for (l = connections; l != NULL; l = l->next)
+    {
+      GDBusConnection *connection = l->data;
+      g_dbus_connection_emit_signal (connection,
+        NULL, g_dbus_interface_skeleton_get_object_path (G_DBUS_INTERFACE_SKELETON (skeleton)), "io.github.kelvinnovais.Database", "DatabaseUpdated",
+        signal_variant, NULL);
+    }
+  g_variant_unref (signal_variant);
+  g_list_free_full (connections, g_object_unref);
+}
+
+static void
+_gawake_server_database_on_signal_rule_canceled (
+    GawakeServerDatabase *object)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (object);
+
+  GList      *connections, *l;
+  GVariant   *signal_variant;
+  connections = g_dbus_interface_skeleton_get_connections (G_DBUS_INTERFACE_SKELETON (skeleton));
+
+  signal_variant = g_variant_ref_sink (g_variant_new ("()"));
+  for (l = connections; l != NULL; l = l->next)
+    {
+      GDBusConnection *connection = l->data;
+      g_dbus_connection_emit_signal (connection,
+        NULL, g_dbus_interface_skeleton_get_object_path (G_DBUS_INTERFACE_SKELETON (skeleton)), "io.github.kelvinnovais.Database", "RuleCanceled",
+        signal_variant, NULL);
+    }
+  g_variant_unref (signal_variant);
+  g_list_free_full (connections, g_object_unref);
+}
+
+static void
+_gawake_server_database_on_signal_schedule_requested (
+    GawakeServerDatabase *object)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (object);
+
+  GList      *connections, *l;
+  GVariant   *signal_variant;
+  connections = g_dbus_interface_skeleton_get_connections (G_DBUS_INTERFACE_SKELETON (skeleton));
+
+  signal_variant = g_variant_ref_sink (g_variant_new ("()"));
+  for (l = connections; l != NULL; l = l->next)
+    {
+      GDBusConnection *connection = l->data;
+      g_dbus_connection_emit_signal (connection,
+        NULL, g_dbus_interface_skeleton_get_object_path (G_DBUS_INTERFACE_SKELETON (skeleton)), "io.github.kelvinnovais.Database", "ScheduleRequested",
+        signal_variant, NULL);
+    }
+  g_variant_unref (signal_variant);
+  g_list_free_full (connections, g_object_unref);
+}
+
+static void gawake_server_database_skeleton_iface_init (GawakeServerDatabaseIface *iface);
+#if GLIB_VERSION_MAX_ALLOWED >= GLIB_VERSION_2_38
+G_DEFINE_TYPE_WITH_CODE (GawakeServerDatabaseSkeleton, gawake_server_database_skeleton, G_TYPE_DBUS_INTERFACE_SKELETON,
+                         G_ADD_PRIVATE (GawakeServerDatabaseSkeleton)
+                         G_IMPLEMENT_INTERFACE (GAWAKE_SERVER_TYPE_DATABASE, gawake_server_database_skeleton_iface_init))
+
+#else
+G_DEFINE_TYPE_WITH_CODE (GawakeServerDatabaseSkeleton, gawake_server_database_skeleton, G_TYPE_DBUS_INTERFACE_SKELETON,
+                         G_IMPLEMENT_INTERFACE (GAWAKE_SERVER_TYPE_DATABASE, gawake_server_database_skeleton_iface_init))
+
+#endif
+static void
+gawake_server_database_skeleton_finalize (GObject *object)
+{
+  GawakeServerDatabaseSkeleton *skeleton = GAWAKE_SERVER_DATABASE_SKELETON (object);
+  g_list_free_full (skeleton->priv->changed_properties, (GDestroyNotify) _changed_property_free);
+  if (skeleton->priv->changed_properties_idle_source != NULL)
+    g_source_destroy (skeleton->priv->changed_properties_idle_source);
+  g_main_context_unref (skeleton->priv->context);
+  g_mutex_clear (&skeleton->priv->lock);
+  G_OBJECT_CLASS (gawake_server_database_skeleton_parent_class)->finalize (object);
+}
+
+static void
+gawake_server_database_skeleton_init (GawakeServerDatabaseSkeleton *skeleton)
+{
+#if GLIB_VERSION_MAX_ALLOWED >= GLIB_VERSION_2_38
+  skeleton->priv = gawake_server_database_skeleton_get_instance_private (skeleton);
+#else
+  skeleton->priv = G_TYPE_INSTANCE_GET_PRIVATE (skeleton, GAWAKE_SERVER_TYPE_DATABASE_SKELETON, GawakeServerDatabaseSkeletonPrivate);
+#endif
+
+  g_mutex_init (&skeleton->priv->lock);
+  skeleton->priv->context = g_main_context_ref_thread_default ();
+}
+
+static void
+gawake_server_database_skeleton_class_init (GawakeServerDatabaseSkeletonClass *klass)
+{
+  GObjectClass *gobject_class;
+  GDBusInterfaceSkeletonClass *skeleton_class;
+
+  gobject_class = G_OBJECT_CLASS (klass);
+  gobject_class->finalize = gawake_server_database_skeleton_finalize;
+
+  skeleton_class = G_DBUS_INTERFACE_SKELETON_CLASS (klass);
+  skeleton_class->get_info = gawake_server_database_skeleton_dbus_interface_get_info;
+  skeleton_class->get_properties = gawake_server_database_skeleton_dbus_interface_get_properties;
+  skeleton_class->flush = gawake_server_database_skeleton_dbus_interface_flush;
+  skeleton_class->get_vtable = gawake_server_database_skeleton_dbus_interface_get_vtable;
+
+#if GLIB_VERSION_MAX_ALLOWED < GLIB_VERSION_2_38
+  g_type_class_add_private (klass, sizeof (GawakeServerDatabaseSkeletonPrivate));
+#endif
+}
+
+static void
+gawake_server_database_skeleton_iface_init (GawakeServerDatabaseIface *iface)
+{
+  iface->database_updated = _gawake_server_database_on_signal_database_updated;
+  iface->rule_canceled = _gawake_server_database_on_signal_rule_canceled;
+  iface->schedule_requested = _gawake_server_database_on_signal_schedule_requested;
+}
+
+/**
+ * gawake_server_database_skeleton_new:
+ *
+ * Creates a skeleton object for the D-Bus interface <link linkend="gdbus-interface-io-github-kelvinnovais-Database.top_of_page">io.github.kelvinnovais.Database</link>.
+ *
+ * Returns: (transfer full) (type GawakeServerDatabaseSkeleton): The skeleton object.
+ */
+GawakeServerDatabase *
+gawake_server_database_skeleton_new (void)
+{
+  return GAWAKE_SERVER_DATABASE (g_object_new (GAWAKE_SERVER_TYPE_DATABASE_SKELETON, NULL));
+}
+
