+++ gawake-cli/src/gawake-dbus-server/meson.build
@@ -1,11 +1,9 @@
 gawake_dbus_server_sources = files(
 	'main.c',
-	'database-connection.c',
+	'dbus-server.c',
 
-	'../dbus-server/dbus-server.c',
-
-	'../utils/gawake-types.c',
-	'../utils/validate-rtcwake-args.c',
 	'../utils/get-time.c',
 	'../utils/debugger.c'
 )
+
+run_target('gdbus-codegenerator', command: 'gdbus-codegenerator.sh')
