+++ gawake-cli/src/gawaked/scheduler.h
@@ -7,8 +7,8 @@
 
 // Structure for the upcoming turn off rule
 typedef struct {
-  gboolean found;
-  gboolean gawake_status;
+  bool found;
+  bool gawake_status;
   int hour;
   int minutes;
   Mode mode;
