+++ gawake-cli/src/utils/database-connection.c
@@ -21,35 +21,37 @@
 #include <sqlite3.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 #include "database-connection.h"
 
-#include "../utils/validate-rtcwake-args.h"
-#include "../utils/debugger.h"
+#include "validate-rtcwake-args.h"
+#include "debugger.h"
+#include "colors.h"
 
 static sqlite3 *db = NULL;
 
 // This function connect to the database
 // Should be called once
-gint connect_database (void)
+int connect_database (void)
 {
   // Open the SQLite database
-  gint rc = sqlite3_open_v2 (DB_NAME, &db, SQLITE_OPEN_READWRITE, NULL);
+  int rc = sqlite3_open_v2 (DB_PATH, &db, SQLITE_OPEN_READWRITE, NULL);
 
   if (rc != SQLITE_OK)
     {
       DEBUG_PRINT_CONTEX;
-      g_fprintf (stderr, "Can't open database: %s\n", sqlite3_errmsg (db));
+      fprintf (stderr, RED ("Can't open database: %s\n"), sqlite3_errmsg (db));
       sqlite3_close (db);
       return EXIT_FAILURE;
     }
   else
     {
-      gchar *err_msg = 0;
+      char *err_msg = 0;
 
-      const gchar UPTDATE_VERSION[] = "UPDATE config SET version = '" VERSION "';";
+      const char UPTDATE_VERSION[] = "UPDATE config SET version = '" VERSION "';";
       rc = sqlite3_exec (db, UPTDATE_VERSION, NULL, 0, &err_msg);
-      g_fprintf (stdout, "Database opened successfully\n");
+      fprintf (stdout, "Database opened successfully\n");
 
       sqlite3_free (err_msg);
     }
@@ -59,26 +61,26 @@ gint connect_database (void)
 
 void close_database (void)
 {
-  g_fprintf (stdout, "Closing database\n");
+  fprintf (stdout, "Closing database\n");
 
   sqlite3_close (db);
 }
 
-static gint validate_rule (const gRule *rule)
+static int validate_rule (const Rule *rule)
 {
   // name can't be bigger than the max value
-  gboolean name = !(strlen (rule->name) >= RULE_NAME_LENGTH);     // (>=) do not include null terminator
+  bool name = !(strlen (rule->name) >= RULE_NAME_LENGTH);     // (>=) do not include null terminator
 
   // hour [00,23]
-  gboolean hour = rule->hour <= 23;     // Due to the data type, it is always >= 0
+  bool hour = rule->hour <= 23;     // Due to the data type, it is always >= 0
 
   // minutes according to enum predefined values
-  gboolean minutes = rule->minutes <= 59;
+  bool minutes = rule->minutes <= 59;
 
   // mode according to enum predefined values
-  gboolean mode = (rule->mode >= 0 && rule->mode <= OFF);
+  bool mode = (rule->mode >= 0 && rule->mode <= OFF);
 
-  gboolean table = (rule->table == T_ON || rule->table == T_OFF);
+  bool table = (rule->table == T_ON || rule->table == T_OFF);
 
   DEBUG_PRINT (("VALIDATION:\nName: %d\n\tName length: %lu\n"\
                 "Hour: %d\nMinutes: %d\nMode: %d\nTable: %d",
@@ -88,48 +90,52 @@ static gint validate_rule (const gRule *rule)
     return EXIT_SUCCESS;
   else
     {
-      g_fprintf (stderr, "Invalid rule values\n\n");
+      fprintf (stderr, "Invalid rule values\n\n");
       return EXIT_FAILURE;
     }
 }
 
-static gint validate_table (const Table table)
+static int validate_table (const Table table)
 {
   if (table == T_ON || table == T_OFF)
     return EXIT_SUCCESS;
 
-  g_fprintf (stderr, "Invalid table\n\n");
+  fprintf (stderr, "Invalid table\n\n");
   return EXIT_FAILURE;
 }
 
-static gboolean run_sql (const gchar *sql)
+static int run_sql (const char *sql)
 {
-  gint rc;
-  gchar *err_msg = 0;
+  int rc;
+  char *err_msg = 0;
 
   DEBUG_PRINT (("Generated SQL:\n\t%s", sql));
 
   rc = sqlite3_exec(db, sql, NULL, 0, &err_msg);
   if (rc != SQLITE_OK)
-    g_fprintf (stderr, "Failed to run SQL: %s\n", sqlite3_errmsg(db));
+    fprintf (stderr, "Failed to run SQL: %s\n", sqlite3_errmsg(db));
 
   sqlite3_free (err_msg);
 
   if (rc == SQLITE_OK)
-    return TRUE;
+    return EXIT_SUCCESS;
   else
-    return FALSE;
+    return EXIT_FAILURE;
 }
 
-gboolean add_rule (const gRule *rule)
+int add_rule (const Rule *rule)
 {
   if (validate_rule (rule))
-    return FALSE;
+    return EXIT_FAILURE;
 
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
-    return FALSE;
+    {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
+    }
 
   switch (rule->table)
     {
@@ -159,25 +165,29 @@ gboolean add_rule (const gRule *rule)
                         rule->mode);
       break;
     default:
-      g_free (sql);
-      return FALSE;
+      free (sql);
+      return EXIT_FAILURE;
     }
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
 
-gboolean edit_rule (const gRule *rule)
+int edit_rule (const Rule *rule)
 {
   if (validate_rule (rule))
-    return FALSE;
+    return EXIT_FAILURE;
 
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
-    return FALSE;
+    {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
+    }
 
   switch (rule->table)
     {
@@ -211,69 +221,76 @@ gboolean edit_rule (const gRule *rule)
                         rule->id);
       break;
     default:
-      g_free (sql);
-      return FALSE;
+      free (sql);
+      return EXIT_FAILURE;
     }
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
 
-gboolean delete_rule (const guint16 id, const Table table)
+int delete_rule (const uint16_t id, const Table table)
 {
   if (validate_table (table))
-    return FALSE;
+    return EXIT_FAILURE;
 
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
-    return FALSE;
+    {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
+    }
 
   sqlite3_snprintf (ALLOC, sql, "DELETE FROM %s WHERE id = %d;", TABLE[table], id);
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
 
-gboolean enable_disable_rule (const guint16 id, const Table table, const gboolean active)
+int enable_disable_rule (const uint16_t id, const Table table, const bool active)
 {
   if (validate_table (table))
-    return FALSE;
+    return EXIT_FAILURE;
 
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
     {
-      g_fprintf (stderr, "Could'n allocate memory\n\n");
-      return FALSE;
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
     }
 
   sqlite3_snprintf (ALLOC, sql, "UPDATE %s SET active = %d WHERE id = %d;", TABLE[table], active, id);
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
 
-gboolean query_rule (const guint16 id, const Table table, gRule *rule)
+int query_rule (const uint16_t id, const Table table, Rule *rule)
 {
   if (validate_table (table))
-    return FALSE;
+    return EXIT_FAILURE;
 
   // Database related variables
-  gint rc;
+  int rc;
   struct sqlite3_stmt *stmt;
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
     {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   // Generate SQL
@@ -285,10 +302,10 @@ gboolean query_rule (const guint16 id, const Table table, gRule *rule)
   if (sqlite3_prepare_v2 (db, sql, -1, &stmt, NULL) == SQLITE_OK
       && sqlite3_step (stmt) != SQLITE_ROW)
     {
-      g_fprintf (stderr, "Invalid ID\n\n");
-      g_free (sql);
+      fprintf (stderr, "Invalid ID\n\n");
+      free (sql);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   // Query data
@@ -300,30 +317,30 @@ gboolean query_rule (const guint16 id, const Table table, gRule *rule)
    *                                                  only for turn off rules
    */
   // Temporary variables to receive the minutes and pass to the structure;
-  gint hour, minutes;
-  gchar timestamp[9]; // HH:MM:SS'\0' = 9 characters
+  int hour, minutes;
+  char timestamp[9]; // HH:MM:SS'\0' = 9 characters
   while ((rc = sqlite3_step (stmt)) == SQLITE_ROW)
     {
       // ID
-      rule->id = (guint16) sqlite3_column_int (stmt, 0);
+      rule->id = (uint16_t) sqlite3_column_int (stmt, 0);
       // NAME
-      g_snprintf (rule->name, RULE_NAME_LENGTH, "%s", sqlite3_column_text (stmt, 1));
+      snprintf (rule->name, RULE_NAME_LENGTH, "%s", sqlite3_column_text (stmt, 1));
 
       // MINUTES AND HOUR
       sqlite3_snprintf (9, timestamp, "%s", sqlite3_column_text (stmt, 2));
       sscanf (timestamp, "%02d:%02d", &hour, &minutes);
-      rule->hour = (guint8) hour;
+      rule->hour = (uint8_t) hour;
       rule->minutes = minutes;
 
       // DAYS
-      for (gint i = 0; i <= 6; i++)
+      for (int i = 0; i <= 6; i++)
         {
           // days range: [0,6]                  column range: [3,9]
-          rule->days[i] = (gboolean) sqlite3_column_int (stmt, (i+3));
+          rule->days[i] = (bool) sqlite3_column_int (stmt, (i+3));
         }
 
       // ACTIVE
-      rule->active = (gboolean) sqlite3_column_int (stmt, 10); // active
+      rule->active = (bool) sqlite3_column_int (stmt, 10); // active
 
       // MODE (for turn on rules it isn't used):
       rule->mode = (Mode) ((table == T_OFF) ? sqlite3_column_int (stmt, 11) : 0);
@@ -334,32 +351,35 @@ gboolean query_rule (const guint16 id, const Table table, gRule *rule)
 
   if (rc != SQLITE_DONE)
     {
-      g_fprintf (stderr, "ERROR (failed to query rule): %s\n", sqlite3_errmsg (db));
-      g_free (sql);
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, "ERROR (failed to query rule): %s\n", sqlite3_errmsg (db));
+      free (sql);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   sqlite3_finalize (stmt);
-  g_free (sql);
+  free (sql);
 
-  return TRUE;
+  return EXIT_SUCCESS;
 }
 
-gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
+int query_rules (const Table table, Rule **rules, uint16_t *rowcount)
 {
   if (validate_table (table))
-    return FALSE;
+    return EXIT_FAILURE;
 
   // Database related variables
-  gint rc;
+  int rc;
   struct sqlite3_stmt *stmt;
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
     {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   // Count the number of rows
@@ -367,23 +387,26 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
   if (sqlite3_prepare_v2 (db, sql, -1, &stmt, NULL) == SQLITE_OK
       && sqlite3_step (stmt) != SQLITE_ROW)
     {
-      g_fprintf (stderr, "ERROR: Failed to query row count\n");
-      g_free (sql);
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to query row count\n"));
+      free (sql);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
   *rowcount = sqlite3_column_int (stmt, 0);
 
   DEBUG_PRINT (("Row count: %d", *rowcount));
 
   // Allocate structure array
+  // https://www.youtube.com/watch?v=lq8tJS3g6tY
   *rules = malloc (*rowcount * sizeof (**rules));
   if (*rules == NULL)
     {
-      g_fprintf (stderr, "ERROR: Failed to allocate memory\n");
-      g_free (sql);
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      free (sql);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   // Generate SQL
@@ -395,13 +418,15 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
   // Prepare statement
   if (sqlite3_prepare_v2 (db, sql, -1, &stmt, NULL) != SQLITE_OK)
     {
-      g_fprintf (stderr, "ERROR: Failed to query rule\n");
-      g_free (sql);
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to query rule\n"));
+      free (sql);
+      free (*rules);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
-  gint counter = 0;
+  int counter = 0;
   // Query data
   /* ATTENTION columns numbers:
    *    0                     1     2             3       4       (...)       10      11        12
@@ -411,8 +436,8 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
    *                                                                      only for turn off rules
    */
   // Temporary variables to receive the hour and minutes, and then pass to the structure
-  gint hour, minutes;
-  gchar timestamp[9]; // HH:MM:SS'\0' = 9 characters
+  int hour, minutes;
+  char timestamp[9]; // HH:MM:SS'\0' = 9 characters
   while ((rc = sqlite3_step (stmt)) == SQLITE_ROW)
     {
       // If the loop tries to assign on non allocated space, leave
@@ -420,20 +445,22 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
         break;
 
       // ID
-      (*rules)[counter].id = (guint16) sqlite3_column_int (stmt, 1);
+      (*rules)[counter].id = (uint16_t) sqlite3_column_int (stmt, 1);
 
       // NAME
       // Allocate memory for the name
-      gint size = sqlite3_column_int (stmt, 0);
-      (*rules)[counter].name = (gchar *) g_malloc (size + 1);
+      int size = sqlite3_column_int (stmt, 0);
+      (*rules)[counter].name = (char *) malloc (size + 1);
       if ((*rules)[counter].name == NULL)
         {
-          g_fprintf (stderr, "ERROR: Failed to allocate memory\n");
-          return FALSE;
+          DEBUG_PRINT_CONTEX;
+          // TODO avoid memory leaking: free names allocated up to here and then *rules
+          fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+          return EXIT_FAILURE;
         }
 
       // Assign name
-      g_snprintf ((*rules)[counter].name,         // string pointer
+      snprintf ((*rules)[counter].name,         // string pointer
                   size + 1,                       // size
                   "%s",                           // format
                   sqlite3_column_text (stmt, 2)   // arguments
@@ -442,18 +469,18 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
       // MINUTES AND HOUR
       sqlite3_snprintf (9, timestamp, "%s", sqlite3_column_text (stmt, 3));
       sscanf (timestamp, "%02d:%02d", &hour, &minutes);
-      (*rules)[counter].hour =  (guint8) hour;
-      (*rules)[counter].minutes = (guint8) minutes;
+      (*rules)[counter].hour =  (uint8_t) hour;
+      (*rules)[counter].minutes = (uint8_t) minutes;
 
       // DAYS
-      for (gint i = 0; i <= 6; i++)
+      for (int i = 0; i <= 6; i++)
         {
           // days range: [0,6]                  column range: [4,10]
-          (*rules)[counter].days[i] = (gboolean) sqlite3_column_int (stmt, (i+4));
+          (*rules)[counter].days[i] = (bool) sqlite3_column_int (stmt, (i+4));
         }
 
       // ACTIVE
-      (*rules)[counter].active = (gboolean) sqlite3_column_int (stmt, 11);
+      (*rules)[counter].active = (bool) sqlite3_column_int (stmt, 11);
 
       // MODE (for turn on rules it isn't used, assigning 0):
       (*rules)[counter].mode = (Mode) ((table == T_OFF) ? sqlite3_column_int (stmt, 12) : 0);
@@ -466,29 +493,35 @@ gboolean query_rules (const Table table, gRule **rules, guint16 *rowcount)
 
   if (rc != SQLITE_DONE)
     {
-      g_fprintf (stderr, "ERROR (failed to query rule): %s\n", sqlite3_errmsg (db));
-      g_free (sql);
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR (failed to query rule): %s\n"), sqlite3_errmsg (db));
+      free (sql);
+      free (*rules);
       sqlite3_finalize (stmt);
-      return FALSE;
+      return EXIT_FAILURE;
     }
 
   sqlite3_finalize(stmt);
-  g_free (sql);
+  free (sql);
 
-  return TRUE;
+  return EXIT_SUCCESS;
 }
 
-gboolean custom_schedule (const guint8 hour,
-                          const guint8 minutes,
-                          const guint8 day,
-                          const guint8 month,
-                          const guint16 year,
-                          const guint8 mode)
+int custom_schedule (const uint8_t hour,
+                          const uint8_t minutes,
+                          const uint8_t day,
+                          const uint8_t month,
+                          const uint16_t year,
+                          const uint8_t mode)
 {
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
-    return FALSE;
+    {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
+    }
 
   RtcwakeArgs rtcwake_args = {
     .hour = hour,
@@ -500,7 +533,7 @@ gboolean custom_schedule (const guint8 hour,
   };
 
   if (validade_rtcwake_args (&rtcwake_args) == -1)
-    return FALSE;
+    return EXIT_FAILURE;
 
   sqlite3_snprintf (ALLOC, sql,
                     "UPDATE custom_schedule "\
@@ -510,20 +543,24 @@ gboolean custom_schedule (const guint8 hour,
                     "WHERE id = 1;",
                     hour, minutes,
                     day, month, year,
-                    mode, TRUE);
+                    mode, true);
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
 
-gboolean schedule (void)
+int schedule (void)
 {
-  gchar *sql = 0;
-  sql = (gchar *) g_malloc (ALLOC);
+  char *sql = 0;
+  sql = (char *) malloc (ALLOC);
   if (sql == NULL)
-    return FALSE;
+    {
+      DEBUG_PRINT_CONTEX;
+      fprintf (stderr, RED ("ERROR: Failed to allocate memory\n"));
+      return EXIT_FAILURE;
+    }
 
   sqlite3_snprintf (ALLOC, sql,
                     "UPDATE custom_schedule "\
@@ -533,10 +570,10 @@ gboolean schedule (void)
                     "WHERE id = 1;",
                     0, 0,
                     0, 0, 0,
-                    0, FALSE);
+                    0, false);
 
-  gboolean ret = run_sql (sql);
-  g_free (sql);
+  int ret = run_sql (sql);
+  free (sql);
 
   return ret;
 }
